Converted short loops
https://releases.llvm.org/13.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/bugprone-too-small-loop-variable.html

+++ utils/RCBot2_meta/bot.cpp
@@ -1946,7 +1946,7 @@ void CBot :: listenForPlayers ()
 
 	m_bListenPositionValid = false;
 
-	for (short int i = 1; i <= gpGlobals->maxClients; i ++ )
+	for (int i = 1; i <= gpGlobals->maxClients; i ++ )
 	{
 		edict_t* pPlayer = INDEXENT(i);
