+++ utils/RCBot2_meta/bot_css_bot.cpp
@@ -230,7 +230,7 @@ void CCSSBot::listenForPlayers()
 
 	m_bListenPositionValid = false;
 
-	for(short int i = 1; i <= gpGlobals->maxClients; i ++)
+	for(int i = 1; i <= gpGlobals->maxClients; i ++)
 	{
 		edict_t* pPlayer = INDEXENT(i);
