+++ utils/RCBot2_meta/bot_getprop.cpp
@@ -711,7 +711,7 @@ edict_t *CClassInterface::FindEntityByNetClassNearest(Vector vstart, const char 
 	edict_t *pfound = NULL;
 	float fMindist = 8192.0f;
 
-	for (short int i = 0; i < gpGlobals->maxEntities; i++)
+	for (int i = 0; i < gpGlobals->maxEntities; i++)
 	{
 		edict_t* current = engine->PEntityOfEntIndex(i);
 		if (current == NULL)
