+++ utils/RCBot2_meta/bot_mods.h
@@ -223,7 +223,7 @@ public:
 	{
 		int count = 0;
 
-		for ( short int i = 0; i < m_iNumControlPoints; i ++ )
+		for ( int i = 0; i < m_iNumControlPoints; i ++ )
 		{
 			if ( m_iOwner[i] == iTeam )
 				count++;
@@ -252,7 +252,7 @@ public:
 	{
 		int count = 0;
 
-		for ( short int i = 0; i < m_iNumControlPoints; i ++ )
+		for ( int i = 0; i < m_iNumControlPoints; i ++ )
 		{
 			if ( canDefendBomb(iTeam,i) )
 				count++;
@@ -265,7 +265,7 @@ public:
 	{
 		int count = 0;
 
-		for ( short int i = 0; i < m_iNumControlPoints; i ++ )
+		for ( int i = 0; i < m_iNumControlPoints; i ++ )
 		{
 			if ( canDefuseBomb(iTeam,i) )
 				count++;
@@ -278,7 +278,7 @@ public:
 	{
 		int count = 0;
 
-		for ( short int i = 0; i < m_iNumControlPoints; i ++ )
+		for ( int i = 0; i < m_iNumControlPoints; i ++ )
 		{
 			if ( canPlantBomb(iTeam,i) )
 				count += getNumBombsRequired(i);
@@ -306,7 +306,7 @@ public:
 	{
 		int count = 0;
 
-		for ( short int i = 0; i < m_iNumControlPoints; i ++ )
+		for ( int i = 0; i < m_iNumControlPoints; i ++ )
 		{
 			if ( canPlantBomb(iTeam,i) )
 				count += getNumBombsRemaining(i);
@@ -329,7 +329,7 @@ public:
 	{
 		int count = 0;
 
-		for ( short int i = 0; i < m_iNumControlPoints; i ++ )
+		for ( int i = 0; i < m_iNumControlPoints; i ++ )
 		{
 			if ( m_iOwner[i] == iTeam )
 				count++;
@@ -451,7 +451,7 @@ public:
 
 	int getFlagID ( edict_t *pent )
 	{
-		for ( short int i = 0; i < m_iNumControlPoints; i ++ )
+		for ( int i = 0; i < m_iNumControlPoints; i ++ )
 		{
 			if ( m_pFlags[i] == pent )
 				return i;
@@ -465,7 +465,7 @@ public:
 		if ( pent == NULL )
 			return -1;
 
-		for ( short int i = 0; i < m_iNumControlPoints; i ++ )
+		for ( int i = 0; i < m_iNumControlPoints; i ++ )
 		{
 			if ( m_pBombs[i][0] == pent || m_pBombs[i][1] == pent )
 				return i;
